fix warnings

+++ include/instruction_code_map.hpp
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 #include <array>
-#include "helpers.hpp"
+#include "helpers/instructionTypes.hpp"
 #include "third_party/unordered_dense.h"
 
 class Instruction;
@@ -15,7 +15,7 @@ struct InstructionInfo;
 extern ankerl::unordered_dense::map<uint32_t, InstructionType_and_addMode> instructionMap;
 
 // generic function pointer for the decode function (it takes a pointer to an instruction, a reference to InstructionInfo struct and an integer as parameters and returns void)
-using DecodeFunc = void (*)(Instruction*, const InstructionInfo&, int);
+using DecodeFunc = void (*)(Instruction*, const InstructionInfo&, uint8_t);
 
 // map for the decode functions (for every key(addressing mode) there is a value(decode function))
 extern const std::array<DecodeFunc, (size_t)AddressingMode::COUNT> addressingModes;
